# use sync.Mutex instead of sync.RWMutex