Apart from CurrentSize and NumItems (which are used by the http status
page), we only lock this mutex for rw, never for reading only. So let's
use the simpler sync.Mutex type.

Also, use the de-facto standard variable name for mutexes in Go code, "mu".