@buchgr 2019-11-26
Good catch. I think we should probably use a different lock for reading and writing though?