@mostynb 2019-11-26
> Good catch. I think we should probably use a different lock for reading and writing though?

CurrentSize and NumItems are the only methods that would be able to use a read-only lock, and they're only called from the http status page. I assume the status page is loaded relatively infrequently so it might be better to just use a simple Mutex, which I assume is slightly faster than RWMutex for write-heavy workloads like we have.

MaxSize acquires the lock but only reads from a field that is never modified, it may as well just read
the field without locking. I will push a new commit to remove it.