@mostynb 2019-11-28
I tried to benchmark sync.Mutex vs sync.RWMutex's write lock to confirm my assumption (RWMutex should be slower in write-heavy scenarios). I used BenchmarkMutexWrite from this stack overflow question:
https://stackoverflow.com/questions/57562606/why-does-sync-mutex-largely-drop-performance-when-goroutine-contention-is-more-t

Then I changed it to use sync.RWMutex instead and compared the results with benchstat. The "old" column here uses sync.Mutex and "new" uses sync.RWMutex:

name                          old time/op  new time/op  delta
MutexWrite/goroutines-2400-8   163ns ± 3%   261ns ± 2%   +59.68%  (p=0.008 n=5+5)
MutexWrite/goroutines-2480-8   165ns ± 3%   273ns ± 3%   +65.45%  (p=0.008 n=5+5)
MutexWrite/goroutines-2560-8   170ns ± 4%   292ns ± 2%   +71.88%  (p=0.008 n=5+5)
MutexWrite/goroutines-2640-8   181ns ± 7%   366ns ± 1%  +102.10%  (p=0.008 n=5+5)
MutexWrite/goroutines-2720-8   317ns ± 2%   372ns ± 1%   +17.34%  (p=0.008 n=5+5)
MutexWrite/goroutines-2800-8   322ns ± 2%   376ns ± 6%   +16.75%  (p=0.008 n=5+5)
MutexWrite/goroutines-2880-8   318ns ± 2%   366ns ± 6%   +15.03%  (p=0.008 n=5+5)
MutexWrite/goroutines-2960-8   318ns ± 2%   370ns ± 3%   +16.50%  (p=0.008 n=5+5)
MutexWrite/goroutines-3040-8   326ns ± 3%   362ns ± 2%   +11.17%  (p=0.008 n=5+5)
MutexWrite/goroutines-3120-8   332ns ± 1%   367ns ± 1%   +10.61%  (p=0.008 n=5+5)
MutexWrite/goroutines-3200-8   339ns ± 1%   373ns ± 1%   +10.15%  (p=0.008 n=5+5)
MutexWrite/goroutines-3280-8   354ns ± 2%   383ns ± 1%    +8.43%  (p=0.008 n=5+5)
MutexWrite/goroutines-3360-8   356ns ± 1%   394ns ± 3%   +10.72%  (p=0.008 n=5+5)
MutexWrite/goroutines-3440-8   369ns ± 2%   411ns ± 2%   +11.33%  (p=0.008 n=5+5)
MutexWrite/goroutines-3520-8   375ns ± 3%   409ns ± 2%    +9.00%  (p=0.008 n=5+5)
MutexWrite/goroutines-3600-8   376ns ± 4%   431ns ± 1%   +14.69%  (p=0.008 n=5+5)
MutexWrite/goroutines-3680-8   383ns ± 6%   421ns ± 1%    +9.86%  (p=0.008 n=5+5)
MutexWrite/goroutines-3760-8   364ns ± 7%   431ns ± 1%   +18.40%  (p=0.008 n=5+5)
MutexWrite/goroutines-3840-8   356ns ± 6%   437ns ± 2%   +22.88%  (p=0.008 n=5+5)
MutexWrite/goroutines-3920-8   357ns ± 2%   428ns ± 1%   +19.76%  (p=0.008 n=5+5)
MutexWrite/goroutines-4000-8   362ns ± 4%   396ns ± 2%    +9.44%  (p=0.008 n=5+5)
MutexWrite/goroutines-4080-8   366ns ± 4%   413ns ± 3%   +13.02%  (p=0.008 n=5+5)
MutexWrite/goroutines-4160-8   371ns ± 4%   411ns ± 3%   +10.61%  (p=0.008 n=5+5)
MutexWrite/goroutines-4240-8   370ns ± 3%   413ns ± 3%   +11.39%  (p=0.008 n=5+5)
MutexWrite/goroutines-4320-8   380ns ± 1%   411ns ± 3%    +8.27%  (p=0.008 n=5+5)
MutexWrite/goroutines-4400-8   392ns ± 2%   421ns ± 6%    +7.29%  (p=0.008 n=5+5)
MutexWrite/goroutines-4480-8   388ns ± 2%   420ns ± 4%    +8.19%  (p=0.008 n=5+5)
MutexWrite/goroutines-4560-8   392ns ± 2%   427ns ± 1%    +8.87%  (p=0.008 n=5+5)
MutexWrite/goroutines-4640-8   398ns ± 1%   431ns ± 2%    +8.24%  (p=0.008 n=5+5)
MutexWrite/goroutines-4720-8   411ns ± 3%   443ns ± 3%    +7.84%  (p=0.008 n=5+5)